+++ src/Serwisant/SerwisantApi/Types/SchemaCustomer/CustomerMutation.php
@@ -42,11 +42,12 @@ class CustomerMutation extends Types\RootType
    * @param RepairInput $repair
    * @param RepairItemInput[] $additionalItems
    * @param string[] $temporaryFiles
+   * @param RepairCreationOptions $options
    * @return RepairCreationResult
    */
-  public function createRepair(RepairInput $repair, array $additionalItems = array(), array $temporaryFiles = array(), $vars = array())
+  public function createRepair(RepairInput $repair, array $additionalItems = array(), array $temporaryFiles = array(), RepairCreationOptions $options = null, $vars = array())
   {
-     return $this->inputArgs('createRepair', array_merge($vars, ['repair' => $repair, 'additionalItems' => $additionalItems, 'temporaryFiles' => $temporaryFiles]));
+     return $this->inputArgs('createRepair', array_merge($vars, ['repair' => $repair, 'additionalItems' => $additionalItems, 'temporaryFiles' => $temporaryFiles, 'options' => $options]));
   }
 
   /**
@@ -62,11 +63,12 @@ class CustomerMutation extends Types\RootType
   /**
    * @param TicketInput $ticket
    * @param string[] $temporaryFiles
+   * @param TicketCreationOptions $options
    * @return TicketCreationResult
    */
-  public function createTicket(TicketInput $ticket, array $temporaryFiles = array(), $vars = array())
+  public function createTicket(TicketInput $ticket, array $temporaryFiles = array(), TicketCreationOptions $options = null, $vars = array())
   {
-     return $this->inputArgs('createTicket', array_merge($vars, ['ticket' => $ticket, 'temporaryFiles' => $temporaryFiles]));
+     return $this->inputArgs('createTicket', array_merge($vars, ['ticket' => $ticket, 'temporaryFiles' => $temporaryFiles, 'options' => $options]));
   }
 
   /**