+++ src/Serwisant/SerwisantApi/Types/SchemaCustomer/Device.php
@@ -1,0 +1,55 @@
+<?php
+
+namespace Serwisant\SerwisantApi\Types\SchemaCustomer;
+
+use Serwisant\SerwisantApi;
+use Serwisant\SerwisantApi\Types;
+
+class Device extends Types\Type
+{
+  /**
+   * @var string
+  */
+  public $ID;
+
+  /**
+   * @var CustomFieldValue[]
+   * Will return a list of values for custom fields
+  */
+  public $customFields;
+
+  /**
+   * @var File[]
+  */
+  public $files;
+
+  /**
+   * @var string
+  */
+  public $model;
+
+  /**
+   * @var string
+  */
+  public $number;
+
+  /**
+   * @var string
+  */
+  public $serial;
+
+  /**
+   * @var Dictionary
+  */
+  public $type;
+
+  /**
+   * @var string
+  */
+  public $vendor;
+
+  protected function schemaNamespace()
+  {
+    return 'SchemaCustomer';
+  }
+}