+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.2.0 - BREAKING CHANGES
+
+- arguments for `createRepair` in customer schema changed
+- arguments for `createTicket` in customer schema changed
+- new queries related to devices in customer schema
+- notes handlinf in service schema
+- repair delegation handling in service schema
+
 ## 3.1.0 - BREAKING CHANGES
 
 - removed `AccessTokenContainerShm` - it's useless and possibly confusing